Haskell build cleanup

+++ opencog/nlp/lojban/CWrapper/build.sh
@@ -1,21 +1,22 @@
-SOURCE_DIR=$1
+BUILD_DIR=$1
 
 export LD_LIBRARY_PATH="${LD_LIBRARY_PATH}:/usr/local/lib/opencog"
 
 libname=$(stack query --allow-different-user | awk 'NR==5' | sed 's/://g'| sed 's/ //g')
 libver=$(stack query --allow-different-user | awk 'NR==7' | sed 's/version: //g' | sed "s/'//g" | sed "s/ //g")
 
-#Cleanup of last build if it exists
-rm -f "$SOURCE_DIR/lib$libname-$libver.so"
 
-#Force Rebuild
-rm -fr ".stack-work/"
+if [ "$(id -u)" -ne 0 ];
+then
+    #Force Rebuild
+    rm -fr ".stack-work/"
 
-# Build haskell bindings package.
-stack build --allow-different-user
+    # Build haskell bindings package.
+    stack build --allow-different-user
+fi
 
 LIB=$(find . -name "*$libname*.so" | awk 'NR==1')
 
 patchelf --set-soname "lib$libname-$libver.so" $LIB
 
-cp $LIB "$SOURCE_DIR/lib$libname-$libver.so"
+cp $LIB "$BUILD_DIR/lib$libname-$libver.so"
